+++ lib/src/analyzers/lint_analyzer/rules/rules_list/avoid_returning_widgets/config_parser.dart
@@ -3,6 +3,7 @@ part of 'avoid_returning_widgets_rule.dart';
 class _ConfigParser {
   static const _ignoredNamesConfig = 'ignored-names';
   static const _ignoredAnnotationsConfig = 'ignored-annotations';
+  static const _allowNullable = 'allow-nullable';
 
   static Iterable<String> getIgnoredNames(Map<String, Object> config) =>
       _getIterable(config, _ignoredNamesConfig) ?? [];
@@ -10,6 +11,9 @@ class _ConfigParser {
   static Iterable<String> getIgnoredAnnotations(Map<String, Object> config) =>
       _getIterable(config, _ignoredAnnotationsConfig) ??
       functionalWidgetAnnotations;
+
+  static bool getAllowNullable(Map<String, Object> config) =>
+      config[_allowNullable] as bool? ?? false;
 
   static Iterable<String>? _getIterable(
     Map<String, Object> config,