+++ test/src/analyzers/lint_analyzer/rules/rules_list/avoid_returning_widgets/examples/nullable_example.dart
@@ -1,0 +1,32 @@
+class AnotherWidget extends StatelessWidget {
+  Widget? get widgetGetter => Container();
+
+  List<Widget>? _getWidgetsList() => [Container()].toList();
+
+  @override
+  Widget build(BuildContext context) {
+    Widget? _localBuildMyWidget() {
+      return Container();
+    }
+
+    _localBuildMyWidget();
+
+    _getWidgetsList();
+
+    return Container();
+  }
+}
+
+Widget? _globalBuildMyWidget() {
+  return Container();
+}
+
+class Widget {}
+
+class Container extends Widget {
+  final Widget? child;
+
+  const Container({this.child});
+}
+
+class StatelessWidget extends Widget {}